% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBNG.to.Lat.Long.R
\name{GBNG.to.Lat.Long}
\alias{GBNG.to.Lat.Long}
\title{GBNG.to.Lat.Long}
\usage{
GBNG.to.Lat.Long(GBNG)
}
\arguments{
\item{GBNG}{Must be a spatial object with the data to plot.}
}
\value{
Latitiude and Longitude
}
\description{
This function take a GB national Grid reference an return Lat Long
}
