% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Green.Carpet.Data.R
\docType{data}
\name{Green.Carpet.Data}
\alias{Green.Carpet.Data}
\title{Green Carpet Data for Leicestershire VC55.}
\format{
A data frame with 2207 rows and 14 variables:
\describe{
  \item{Code}{Bradley Fletcher Code}
  \item{Taxon}{Taxon name}
  \item{Date}{Date of observation in source format}
  \item{Gridref}{GB grid reference in source format}
  \item{Stage}{Stage of lifecycle. In this case all records adult}
  \item{YYMMDD}{Date as YYYY-MM-DD}
  \item{week}{Date as week number}
  \item{YYYY}{Date as year}
  \item{DD}{Date as day of month}
  \item{MM}{Date as month}
  \item{MMDD}{Date as MMDD}
  \item{DOY}{Date as day of year}
  \item{Long}{Longitude as WSGS84}
  \item{Lat}{Latitude as WSGS84}

}
}
\usage{
Green.Carpet.Data
}
\description{
A dataset containing 2207 rows of Green Carpet observations.
.
}
\keyword{datasets}
