% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lat.Long.to.GBNG.R
\name{Lat.Long.to.GBNG}
\alias{Lat.Long.to.GBNG}
\title{Lat. Long to GBNG}
\usage{
Lat.Long.to.GBNG(Lat, Long)
}
\arguments{
\item{Lat}{Vector list}

\item{Long}{Vector list}
}
\value{
GBNG with letters
}
\description{
This function takes lat long vectors and output format UKNG
}
