% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/northamptonshire.R
\docType{data}
\name{northamptonshire}
\alias{northamptonshire}
\title{Outline Map of Northamptonshire}
\format{S4 Map object SpatialPolygonsDataFram
\describe{
  \item{NAMES}{Factor} Adjoining county
  \item{REGION}{Factor} Single value in this case: England.

}}
\source{
\url{https://nbn.org.uk/tools-and-resources/nbn-toolbox/watsonian-vice-county-boundaries/}
}
\usage{
northamptonshire
}
\description{
A dataset containing an outline map of Northamptonshire
}
\section{Warning}{


Objects of this type need to use the rgeos plot function. This can be automatically selected
when the library is loaded. Try a simple plot to make sure your system recognises this
of object.
}

\keyword{datasets}
