% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenology.Plot.R
\name{Phenology.Plot}
\alias{Phenology.Plot}
\title{Phenology.Plot}
\usage{
Phenology.Plot(taxon_name, taxon_data)
}
\arguments{
\item{taxon_name}{Name used as the taxon the label the chart.}

\item{taxon_data}{A vector of day of year numbers. Can be strings.}
}
\value{
A chart object
}
\description{
Plots a phenology chart for a taxon. If more than one point is supplied
the a circular density plot is generated
}
