% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.Maps.ggplot.R
\name{Plot.Maps.ggplot}
\alias{Plot.Maps.ggplot}
\title{Plot Maps Density ggPlot}
\usage{
Plot.Maps.ggplot(taxon_name, taxon_spatial_data, map)
}
\arguments{
\item{taxon_name}{Name of taxon used in map description.}

\item{taxon_spatial_data}{Must be a spatial object with the data to plot.}

\item{map}{A map used as the underlay for the plot.}
}
\value{
A Density Map
}
\description{
Plot Maps Density ggPlot
}
