% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.Maps.R
\name{Plot.Maps}
\alias{Plot.Maps}
\title{Plot Maps}
\usage{
Plot.Maps(taxon_name, taxon_spatial_data, map)
}
\arguments{
\item{taxon_name}{Name of taxon used in map description.}

\item{taxon_spatial_data}{Must be a spatial object with the data to plot.}

\item{map}{A map used as the underlay for the plot.}
}
\value{
A map
}
\description{
Plot Maps
}
